topotests: fix duplicate announcements at restart
After pyrtr restart, the RPKI client re-connects and sends the
session ID from the previous session. pyrtr resends it all the prefixes.
However, the client kept the prefixes in cache and does not expect the
server to resend them. Its sends the following error packet (output from
Wireshark):

> RPKI-Router Protocol (Error Report)
>     Version: 0
>     PDU Type: Error Report (10)
>     Error Code: Duplicate Announcement Received (7)
>     Length: 36
>     Length of Encapsulated PDU: 20
>     Erroneous PDU: 000400000000001401181800c63364000000fffa
>     Length of text: 0
>     Erroneous Text:

Set the last_serial to the received serial if the client sends a valid
session ID.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>